# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Savory AI Server API
  description: |
    REST API сервер для системы управления ресторанами Savory AI.

    ## Основные возможности:
    - Управление пользователями и организациями
    - Управление ресторанами и столами
    - Меню и блюда с категориями
    - Бронирование столиков (публичное API для AI-бота и защищённое для персонала)
    - Система подписок
    - AI-чат с поддержкой бронирования через Anthropic Claude
    - Генерация QR-кодов
    - Административная панель

    ## Аутентификация
    Для доступа к защищённым эндпоинтам используется JWT Bearer токен.
    Получите токен через `/auth/login` и передавайте его в заголовке `Authorization: Bearer <token>`.
  version: 1.0.0
  contact:
    name: Savory AI Team
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.savory-ai.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен авторизации. Получите через POST /auth/login
  schemas:
...
    Response:
      type: object
      description: Стандартная обёртка для всех ответов API
      properties:
        code:
          type: integer
          description: HTTP код ответа
          example: 200
        messages:
          type: array
          description: Массив сообщений (успех/ошибки)
          items:
            type: string
          example: ["success"]
        data:
          type: object
          description: Данные ответа
        meta:
          type: object
          description: Метаданные (пагинация и т.д.)
...
    DishResponse:
      type: object
      description: Информация о блюде
      properties:
        id:
          type: integer
        created_at:
          type: string
          format: date-time
        restaurant:
          type: object
          description: Ресторан-владелец
          properties:
            id:
              type: integer
            name:
              type: string
        menu_category:
          type: object
          description: Категория меню
          properties:
            id:
              type: integer
            name:
              type: string
        name:
          type: string
          description: Название блюда
        price:
          type: number
          format: float
          description: Цена
        description:
          type: string
          description: Описание блюда
        image:
          type: string
          description: URL изображения
        ingredients:
          type: array
          description: Список ингредиентов
          items:
            $ref: '#/components/schemas/IngredientResponse'
        allergens:
          type: array
          description: Список аллергенов
          items:
            $ref: '#/components/schemas/AllergenResponse'
    IngredientResponse:
      type: object
      description: Ингредиент блюда
      properties:
        id:
          type: integer
        name:
          type: string
          description: Название ингредиента
        quantity:
          type: number
          format: float
          description: Количество
    AllergenResponse:
      type: object
      description: Аллерген блюда
      properties:
        id:
          type: integer
        name:
          type: string
          description: Название аллергена
        description:
          type: string
          description: Описание аллергена
...
paths:
...
  /dishes/dish-of-day/{restaurant_id}:
    get:
      tags:
        - Dish
      summary: Получить блюдо дня
      description: |
        Возвращает текущее блюдо дня указанного ресторана.
        **Публичный эндпоинт** - не требует авторизации.
      operationId: getDishOfDay
      parameters:
        - name: restaurant_id
          in: path
          required: true
          description: ID ресторана
          schema:
            type: integer
      responses:
        '200':
          description: Блюдо дня
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DishResponse'
        '404':
          description: Блюдо дня не установлено
  /dishes/dish-of-day/{id}:
    post:
      tags:
        - Dish
      summary: Установить блюдо дня
      description: Устанавливает указанное блюдо как блюдо дня.
      operationId: setDishOfDay
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID блюда для установки как блюдо дня
          schema:
            type: integer
      responses:
        '200':
          description: Блюдо дня установлено
        '404':
          description: Блюдо не найдено

  # ==================== Questions ====================
...